+++ example/src/App.js
@@ -24,7 +24,15 @@ export default class App extends Component {
 `;
 const example2 = `<EdiText
   type='textarea'
-  inputClassName="textarea"
+  inputProps={{
+    className: 'textarea',
+    placeholder: 'Type your content here',
+    style: {
+      outline: 'none',
+      minWidth: 'auto'
+    },
+    rows: 5
+  }}
   value="How do you define real? If you're talking about what you can feel, what you can smell,\\
   what you can taste and see, then real is simply electrical signals interpreted by your brain"
   onSave={this.onSave}
@@ -33,7 +41,6 @@ const example2 = `<EdiText
 `;
 const example3 = `<EdiText
   type="textarea"
-  inputClassName="text"
   saveButtonText="Apply"
   cancelButtonText="Cancel"
   editButtonText="Edit"
@@ -49,18 +56,93 @@ const example4 = `<EdiText
   onSave={this.onSave}
 />`
 
-const example5 = `<EdiText
+const example5 = `export default class App extends Component {
+  onSave = val => {
+    console.log('Edited Value -> ', val)
+  }
+
+  validationFailed = textValue => {
+    alert(\`The text <\${textValue}> is not valid.\nYou shall not use the word SMITH here!!!\`)
+  }
+
+  render () {
+    return (
+      <EdiText
+        type="text"
+        validation={val => val.toLowerCase().indexOf('smith') < 0}
+        onValidationFail={this.validationFailed}
+        inputProps={{
+          placeholder: "Don't use the word 'Smith'..."
+        }}
+        value="Why Mr. Anderson? Why? Why? Why?"
+        onSave={this.onSave}
+      />
+    )
+  }
+}
+`
+
+const example6 = `<EdiText
   type="text"
-  className="my-react-header"
+  viewProps={{
+    className: 'my-react-header',
+    style: { borderRadius: 3 }
+  }}
   value="Hello React!"
   onSave={this.onSave}
 />`
 
+const example7 = `<EdiText
+  type="text"
+  hint="It is from Matrix Revolutions."
+  inputProps={{
+    placeholder: 'Type your answer here',
+    style: {
+      backgroundColor: '#233C51',
+      color: '#E6ECF1',
+      fontWeight: 500,
+      width: 250
+    },
+    name: 'answer1'
+  }}
+  viewProps={{
+    className: 'custom-view-class'
+  }}
+  value="No. The honor is still mine."
+  onSave={this.onSave}
+/>`
+
+const example8 = `<EdiText
+  type="text"
+  hint="React is not a framework, it is a library."
+  viewProps={{
+    className: 'react-answer-1',
+    style: { borderRadius: 3 }
+  }}
+  value="I am not sure..."
+  onSave={this.onSave}
+/>`
+
+const example9 = `<EdiText
+  type="date"
+  inputProps={{
+    min: "2000-01-01",
+    max: "2049-01-01"
+  }}
+  hint="All dates are allowed between 2000 and 2049"
+  value={new Date().toDateString()}
+  onSave={this.onSave}
+/>`
 
 export default class App extends Component {
   onSave = val => {
     console.log("Edited Value -> ", val);
-  };
+  }
+
+  validationFailed = textValue => {
+    alert(`The text <${textValue}> is not valid.\nYou shall not use the word SMITH here!!!`)
+  }
+
   render() {
     return (
       <div>
@@ -74,7 +156,7 @@ export default class App extends Component {
                 <span style={{marginRight: 5}}><a className="github-button" href="https://github.com/alioguzhan/react-editext" data-size="large"
                 data-show-count="true" aria-label="Star alioguzhan/react-editext on GitHub">Star</a></span>
                 <span>
-                <a className="github-button" href="https://github.com/alioguzhan/react-editext/fork" data-size="large" 
+                <a className="github-button" href="https://github.com/alioguzhan/react-editext/fork" data-size="large"
                 data-show-count="true" aria-label="Fork alioguzhan/react-editext on GitHub">Fork</a></span>
             </div>
           </div>
@@ -82,7 +164,7 @@ export default class App extends Component {
         <div className="container">
           <div className="tile is-parent is-vertical is-12">
             <p><strong>EdiText</strong> is a React component that converts your text into editable content. Below you can see some examples.</p>
-            <p>For installation and available props please see the 
+            <p>For installation and available props please see the
               <a href="https://github.com/alioguzhan/react-editext"> repository page</a>.
             </p>
           </div>
@@ -108,10 +190,38 @@ export default class App extends Component {
             </div>
           </div>
           <div className="tile is-parent is-vertical is-10">
+            <div className="subtitle">Show a Hint Message</div>
+            <p className="content">
+            If you don't want to use another library and just want to show a simple message as hint, use the <code>hint</code> prop.
+            </p>
+            <div className="columns">
+              <div className="column is-half">
+                <SyntaxHighlighter language="javascript" style={light}>
+                  {example8}
+                </SyntaxHighlighter>
+              </div>
+              <div className="column">
+                <div className="subtitle">Output</div>
+                <p>What is the main difference between React and Angular?</p>
+                <div className="custom-wrapper">
+                  <EdiText
+                    type="text"
+                    hint="React is not a framework, it is a library."
+                    viewProps={{
+                      className: 'react-answer-1',
+                    }}
+                    value="I am not sure..."
+                    onSave={this.onSave}
+                  />
+                </div>
+              </div>
+            </div>
+          </div>
+          <div className="tile is-parent is-vertical is-10">
             <div className="subtitle">Textarea for Editing Input</div>
             <p className="content">
-              You can use {"<textarea>"} for editing if your content is too
-              long. Also you can specify custom class name for the input element.
+              You can use <code>{"<textarea>"}</code> for editing if your content is too
+              long. Also you can specify custom class name and other props for the input element.
               Since this page has built with Bulma.io, we passed <code>textarea</code>
               as class name.
             </p>
@@ -125,7 +235,15 @@ export default class App extends Component {
                 <div className="subtitle">Output</div>
                 <EdiText
                   type="textarea"
-                  inputClassName="textarea"
+                  inputProps={{
+                    className: 'textarea',
+                    placeholder: 'Type your content here',
+                    style: {
+                      outline: 'none',
+                      minWidth: 'auto'
+                    },
+                    rows: 5
+                  }}
                   value="How do you define real? If you're talking about what you can feel, what you can smell,
                   what you can taste and see, then real is simply electrical signals interpreted by your brain"
                   onSave={this.onSave}
@@ -159,6 +277,72 @@ export default class App extends Component {
             </div>
           </div>
           <div className="tile is-parent is-vertical is-10">
+            <div className="subtitle">Custom Props for Input Element</div>
+            <p className="content">
+            It is likely that you will want to add some props to input element.
+            You can pass any valid HTML attribute to the input element via <code>inputProps</code>.
+            </p>
+            <div className="columns">
+              <div className="column is-half">
+                <SyntaxHighlighter language="javascript" style={light}>
+                  {example7}
+                </SyntaxHighlighter>
+              </div>
+              <div className="column">
+                <div className="subtitle">Output</div>
+                <p>What was the Morpheus' answer when Neo say 'It was an honor, sir' ?</p>
+                <div className="custom-wrapper">
+                  <EdiText
+                    type="text"
+                    hint="It is from Matrix Revolutions."
+                    inputProps={{
+                      placeholder: 'Type your answer here',
+                      style: {
+                        backgroundColor: '#233C51',
+                        color: '#E6ECF1',
+                        fontWeight: 500,
+                        width: 250
+                      },
+                      name: 'answer1'
+                    }}
+                    viewProps={{
+                      className: 'custom-view-class'
+                    }}
+                    value="No. The honor is still mine."
+                    onSave={this.onSave}
+                  />
+                </div>
+              </div>
+            </div>
+          </div>
+          <div className="tile is-parent is-vertical is-10">
+            <div className="subtitle">Custom Props for Text Element</div>
+            <p className="content">
+            You may want to add some styling to your text content instead of render it as a plain text. To do that you can use
+            <code>viewProps</code> prop. You can pass any valid HTML attribute.
+            These props will be passed to content <code>div</code>.
+            </p>
+            <div className="columns">
+              <div className="column is-half">
+                <SyntaxHighlighter language="javascript" style={light}>
+                  {example6}
+                </SyntaxHighlighter>
+              </div>
+              <div className="column">
+                <div className="subtitle">Output</div>
+                <EdiText
+                  type="text"
+                  viewProps={{
+                    className: 'my-react-header',
+                    style: { borderRadius: 3 }
+                  }}
+                  value="Hello React!"
+                  onSave={this.onSave}
+                />
+              </div>
+            </div>
+          </div>
+          <div className="tile is-parent is-vertical is-10">
             <div className="subtitle">Validate the Content</div>
             <p className="content">
               You can validate the value of input before save. Just pass a function to <code>validation</code> prop.
@@ -184,11 +368,12 @@ export default class App extends Component {
             </div>
           </div>
           <div className="tile is-parent is-vertical is-10">
-            <div className="subtitle">Custom Class Name for Text</div>
+            <div className="subtitle">Custom Validation</div>
             <p className="content">
-            You may want to add some styling to your text content instad of render it as a plain text. To do that, 
-            you can use <code>className</code> prop. The class name that passed to this prop will be used for styling
-            text.
+              EdiText has a very basic validation handling. You will probably want to use an external library to
+              show custom and fancy validation warnings or messages. You can track the validity of input value with
+              <code>onValidationFail</code> prop. Just pass your method to listen and react against validation updates.
+              <strong>Type the word 'Smith' into below input and try to save it.</strong>
             </p>
             <div className="columns">
               <div className="column is-half">
@@ -200,8 +385,39 @@ export default class App extends Component {
                 <div className="subtitle">Output</div>
                 <EdiText
                   type="text"
-                  className="my-react-header"
-                  value="Hello React!"
+                  validation={val => val.toLowerCase().indexOf('smith') < 0}
+                  onValidationFail={this.validationFailed}
+                  inputProps={{
+                    placeholder: "Don't use the word 'Smith'..."
+                  }}
+                  value="Why Mr. Anderson? Why? Why? Why?"
+                  onSave={this.onSave}
+                />
+              </div>
+            </div>
+          </div>
+          <div className="tile is-parent is-vertical is-10">
+            <div className="subtitle">Usage for Date Values</div>
+            <p className="content">
+              You can use EdiText for date inputs as well. Just type <code>date</code> to type prop.
+              Other related and types are <code>datetime-local</code>, <code>time</code>, <code>week</code>, <code>month</code>
+            </p>
+            <div className="columns">
+              <div className="column is-half">
+                <SyntaxHighlighter language="javascript" style={light}>
+                  {example9}
+                </SyntaxHighlighter>
+              </div>
+              <div className="column">
+                <div className="subtitle">Output</div>
+                <EdiText
+                  type="date"
+                  inputProps={{
+                    min: "2000-01-01",
+                    max: "2049-01-01"
+                  }}
+                  hint="All dates are allowed between 2000 and 2049"
+                  value={new Date().toLocaleString()}
                   onSave={this.onSave}
                 />
               </div>